Fix timestamp issue
In the docs it is written, that adding timestamps to url can be disabled by adding crossOrigin to cropper's img. But it didn't work, cause in load instead of checking that crossOrigin is not set, you checked whether is checked... so timestamp was always added. This is a fix for that, making it possible to disable adding timestamp.