perf: don't recompute the DataUrl if it is already known
This change improves performance when a JPEG image is loaded from a
data url and checkOrientation is set to true. Before this change, the
image was converted into a byte array and then it was converted back
into a data url, without being modified between the two steps. After
this change, we will just reuse the original data url instead of
converting the byte array back. (This saves ~1secs on my laptop with
large photos.)