Improve Accessibility
Improving accessibility of the image to be cropped.  Sets the alt text from the original img tag and if not provided creates a default value that should make sense.  Added tab order so that it can be keyboard accessible and to allow screen readers to read the alt text otherwise the screen reader doesn't announce the image.